pyproject.toml
[project]
name = "python-development-environment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "valiantlynx", email = "gormerykombo@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "debugpy>=1.8.18",
    "fastapi>=0.124.4",
    "loguru>=0.7.3",
    "prometheus-fastapi-instrumentator>=7.1.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.38.0",
]

[project.scripts]
python_development_environment = "python_development_environment:main"

[build-system]
requires = ["uv_build>=0.8.17,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
    "nbstripout>=0.8.2",
]
